feat: 6549 AC (stack version)
Used stack to maintain the leftmost reachable position of each fence.
Stack guarantees that heights of fences in the stack is in increasing order (to top).
Until the top of the stack has smaller height than the current one, we
can pop from the stack to find the leftmost reachable position.
When popping, the current position is the rightmost reachable position
since heights are in increasing order.