feat: 2263 AC
Tree traversing
For a sequence of in-order and post-order for a single subtree,
the last node of the post-order sequence is the root node of the subtree.
In the in-order sequence, the subsequence before the root node is the
sequence for the left subtree, and the subsequence after the root node
is the sequence for the right subtree.
Since the length of subsequences of either subtree in the in-order and
the post-order sequences should be same, we can divide the sequence into
two parts.